System Tray Prototype
This is the first prototype, that implements the top level user
interaction. It creates a system tray icon, and let's the user switch
between on/off states on left click - which will be combined with
tracking. Show the config, project manager and analytics menu options on
right click (only placeholder actions so far). And then the middle click
shows a list of projects.

Note: The final application will have a radio button selection project
list, which I did not add in this prototype yet.