Single Instance Prototype
Let's Code Episode 5 - Source Code. The video for this commit can be
found at: http://youtu.be/eLCWgRnNZao

This prototype tries to create a system wide, named shared memory block
on startup. If it succeeds, it's a primary instance and shows a label
widget. Otherwise it checks the return code and if it determines that it
is a secondary instance, it shows a message box and then quits.